refactor: replace Page with manageError HOC & outsource connect
